Initial prototype support for PEP 735
Introduce a new option, `--dependency-group GROUP` for `pip install`
and `pip download` commands.

`--dependency-group` is an `append` option (can be passed multiple
times) and is resolved via a simple process:
- the requested groups are resolved to lists of PEP 508
  specifiers, using an implementation based on the reference
  implementation in the PEP
- those specifiers are parsed with `install_req_from_req_string`

Dependency groups are parsed by `ReqCommand` internally, and
therefore `pip wheel` provides `options.dependency_groups = []` to
be compatible with its base class, but without exposing this option.

For this initial implementation, `pyproject.toml` is discovered from
the current working directory, and there is no option to control where
it is found.

The signature of `parse_dependency_group` is matched to the
requirements file parser, even though the session and other arguments
are currently ignored.